[WIP] Test: remove runAllTests from superclass
This was initially done so that the setup and teardown methods could be called automatically for all tests, instead of manually for each test. In the end, that idea was abandoned, but the structure stayed. Swift on Linux seems to have a problem with overriding methods defined in extensions, so it suffices to remove it.
Also: Add documentation comments to both runAllTests and allTests, which should clarify their purpose and their differences.