[WIP] Bugfix: detect all errors in Bootstrap unit tests
Errors in the bootstrapped unit tests cause an exception to be printed (to mark the stack trace and tell us where the test failure came from). They should also mark the test suite as a failure, but that wasn't being done in some cases. This should allow us to detect thrown exceptions to make sure they aren't ignored.