[WIP] Feature: print whole results when diffs time out
Calculating diffs for failed test results involves calling the `diff` command line tool, which can be slow and time out (particularly on Docker). In those cases, nothing would be printed and we would have no clue as to what went wrong. Now, when diffs time out, the whole result will be printed, which allows for a manual inspection (or a manual call to diff).