[WIP] Test: add a few missing tests to the test lists
For some reason, these tests were not on the Linux or Kotlin test lists (I probably forgot to add them at some point). Now, all tests should be in these lists, including ones that shouldn't be run (in comments) to allow for an easier inspection (as well as making clear the intention to not run them vs. just forgetting to add them).