Refactor: rename "Test Files" to "Test cases"
The name Test Files might cause confusion with the "Tests" folder as to what are the contents of each (i.e. which one contains the source files for the test suite, and what does the other folder contain). Since the Tests folder was created by the package manager, it's probably best to leave it as it is. The "Test Files" folder is now renamed as "Test cases" which will hopefully be clearer.
Some variables and properties that were found along the way were also renamed accordingly.