Docs: clarify the purpose of errors.swift
This test case was created to purposely trigger more than one error in the translation. It is used by DriverTest.swift to test if Gryphon is capable of continuing after the first error and if it's capable of stopping after the first error. Therefore, if Gryphon ever started supporting the expressions that cause the errors in this file, the person that fixed that might think the errors.swift file was no longer needed. This comment should clarify that.