Refactor: finish centralizing hard coded file paths
This also moves the kotlin build folders to inside the ".gryphon" folder (before they were on the project root). This means less clutter for the user and the possibility of running `gryphon clean` to take care of everything. Also, since the kotlin build folders are now inside ".gryphon", they no longer have to be hidden.

The Kotlin build folder path appeared in a public API and because of that it had to be made public. For consistency, other paths were also made public.