Cleanup: delete "Example ASTs" folder
The "Example ASTs" folder no longer contained ASTs, just the test.swift file. Moving the file to the root of the project might make things less confusing.
I considered deleting the file altogether (that is, removing it from the repo), but it was still useful to have its path hard coded in preBuildScript.sh, and this might make it easier to write a "Getting Started" text.