Feature: start correctly translating access modifiers
So far, the way access modifiers were being translated was simple and worked for Gryphon's limited purposes. However, it could cause cases where members were accessible in Kotlin when they shouldn't be, and vice-versa. This new translation method should better approximate Kotlin's behavior to Swift's, though the documentation for both languages is less than perfect so it's hard to be sure. This commit implements translations for variables in functions inside classes and enums.

Also:
- Add support for access modifiers in variable translations (for translating properties with access modifiers).
- Fix some access modifiers that couldn't be correctly translated into Kotlin.
- Made TranspilationPass.parent into an optional so we can test for it in top-level nodes without crashing.