Bugfix: avoid triggering an invalid index when decoding the AST dump
Some identifiers contain a space in "bla.foo extension.bar". So far this was handled by looking ahead of the space to see if there was an extension, which was done in an unsafe way because of a bug in the Swift compiler that may cause the "extension" string to have a newline in the middle of it.
This commit refactors the ASTDumpDecoder to strip all newlines from the AST dump when it is initialized. This means the algorithm for detecting the " extension" becomes simpler and more correct, and several other algorithms also end up becoming simpler. The newlines were only really used to identify the end of a list of identifiers, so that's the only algorithm that had to be changed.