Bugfix: remove calls to toMutableList from template translations
Kotlin declarations of a few of these methods in List return Lists, not MutableLists. The methods in GryphonSwiftLibrary reflect this, but the translations didn't yet. Fixing this caused a few problems: apparently, Swift detects when a MutableList is expected and uses other methods instead of the ones implemented in the GryphonSwiftLibrary (probably methods from Collection or something). This means we have to call toMutableList manually, which reflects the model used in Kotlin. A tip for detecting when this happens is the color used for syntax highlights in the name of the method (if it's a project color or a language color), otherwise it should cause a Kotlin compilation error.