Feature: finish adding support for access modifiers from annotations
This adds support for annotations in classes, including support for access modifiers in annotations. This should be the last part of supporting access modifiers in annotations.
Unfortunately, it doesn't seem like there's a way to test annotations that aren't access modifiers currently, so there were no tests added for that, only a TODO.