Feature: Start unifying open/final treatment
So far, classes default to being open but their properties and methods do not. The new behavior will be that everything is open by default, unless the user passes a `-default-final` option that says otherwise. In any case, the user should get to decide on a case-by-case basis if each declaration is open or final using annotations.
This starts laying the groundwork for that change. Classes now know if they're open or not, so that we can extract manual changes from annotations. SwiftTranslators now use contexts so we can pass the default option to them. Classes also use annotations as an array, which other nodes will soon be using too - this should make it easier for us to change the array, like we do here and with access modifiers.
There are no new tests yet.