Test: add tests for open/final and for -default-final
This test introduces a new feature to the test suite. Test cases that end in "-default-final" will be compiled with that flag set. This allows us to easily test the -default-final feature in integration and acceptance tests, and to use it in other test cases later if we want to.

Also fix a bug: static declarations are marked as final by Swift only in classes, but not in structs or enums. They have to be manually marked as final in the transpilation pass. The only way to change their final status is with manual annotations, since no other rules apply.