Feature: remove build and run options
These options were more useful in the beginning, when the program was more focused on translating scripts than on Xcode integration. Now, it's worth removing them to make the codebase simpler. It also falls in line with the Unix idea of composability and programs that do only one thing well.

- Add a platform-specific TestUtilities file to contain the OS information, which is now only useful for tests. Rename the old TestUtilities to SharedTestUtilities.
- The kotlinBuild* folder was moved to inside the .gryphon folder a while back, so we can remove it from the gitignore.
- Make the AcceptanceTest create the kotlinBuildFolder if needed, and make it check for the Kotlin compiler instead of the CompileTest.
- Add messages for some DriverTests to facilitate debugging.