Feature: improve collection casting
Casting collections before was done in a way that generated incorrect Kotlin code. The casts in Kotlin weren't checked, since Kotlin doesn't retain information on generic types at runtime. This means casts would always silently succeed but could cause the program to crash when an element was accessed.
The new Kotlin algorithms implement what seems to be the Swift behavior of checking every element before succeeding the cast, which should fix this problem.