Bugfix: fix deadlocks in the Swift shell
Swift's Pipe type implements a pipe with a buffer for storing data. When that buffer fills up, the process gets blocked waiting for it to empty so it can continue running. This means we have to empty the two pipes regularly and asynchronously, otherwise the program will deadlock.

With this bug fixed, there's no longer a need to set a timeout for processes.