Feature: Add support for choosing toolchains
This is the first step towards fully supporting more than one Swift version. It allows us to decide which Swift version to use when dumping the ASTs and to check if the version being used is supported.

It also explicitly tells the compiler to use 5.1 everywhere for now, while 5.2 support is still missing. Once that's done we'll probably be able to remove most of these.