Bugfix: Xcode support needs --toolchain when calling swiftc
It's not enough to call xcodebuild using its toolchain argument. This sets the TOOLCHAIN variable (and it isn't clear yet what that variable does) but it does not set the toolchain when calling swiftc. Instead, it simply calls swiftc from the current toolchain using an absolute path to the executable inside the toolchain.
In order to avoid having to look for the executable for swiftc in each toolchain, we choose to replace it with a call to xcrun so that we can use xcrun's toolchain argument.