Refactor: make the toolchain system more functional
The system that handled toolchains and Swift versions was comprised mainly of global variables. This made it difficult to handle multiple Swift versions one after another without restarting the program (which is needed for testing old Swift versions). Toolchains and Swift versions are now a part of the transpilation context.

Also renames `defaultFinal` to `defaultsToFinal`.