Feature: improve AST printing
ASTs in errors are now printed only when requested by the user with `-print-ASTs-on-error`. ASTs for outputs (i.e. the SwiftAST and rawAST) are printed normally when requested with `-emit-swiftAST` etc.

The `-line-limit=<N>` argument now applies correctly and globally to all AST printing (both on errors and on outputs).