Bugfix: fix the error mapping algorithm
The error mapping was often reporting Kotlin errors in the wrong Swift source locations. This was apparently caused by the mapping scripts that compared the source location incorrectly. They often said a location did not belong in a range even though it did, which caused locations to be matched to the wrong ranges.