Bugfix: the compile Kotlin script stopped when there was a Kotlin error
Kotlin errors caused the gradle call to fail, which caused the whole script to fail and not report the errors for Xcode. The gradle call is now the only one allowed to fail in the script.

Also remove specific checks for failure from the end of the script, since they were redundant.