Bugfix: fix templates with variables that are substrings of others
When replacing a variable's translation in a template string, the order of the variables was random (it was dependent on hash values from the Dictionary implementation). Because of this, when a template had a variable which was a substring of another, it was possible for that variable to be replaced incorrectly.

For instance, say there was a template matching _a to foo and _ab to bar. If _a came first on the dictionary, it was possible to replace _ab as foob instead of bar.

Now, the matches are being ordered by length (decreasing), which should fix this problem.