Feature: add the simulator option
In computers that don't have an iOS device configured, xcodebuild may fail with a code signing error. In those cases, it might be enough to try to make xcodebuild use a simulator. The `--simulator` flag allows the user to choose an installed simulator version to use. The error message that suggests this option also tries to find and suggest an installed iOS version.