Bugfix: failure to setup the project during AST dump
When the AST dump update fails, the compiler tries to setup the project again, to maybe update any missing files. That process can fail, which would result in a message about failing to initialize instead of failing to update the AST dumps. Now, that process fails silently, so that the AST dump message is reported.