Cleanup: remove unused flag and rename old method
The --simulator flag is now being used internally automatically. It should only be needed externally to override the simulator version or to avoid an extra xcodebuild call, but that doesn't seem to be necessary for now.
Also changes the name of the setupGryphonFolder method, since that task is being almost entirely done by the `initialize` method.