Meta: add a script to make it easier to run tests
This adds a script that performs all of the setup needed to run the test suite. It should be runnable from a clean clone of the project. It automatically runs all test files except for Acceptance, Bootstrap and Performance. Acceptance and Bootstrap can be run at the end of the list with `-a` and `-b`, respectively. Running Bootstrap includes calling prepareForBootstrapTests, etc. Performance tests should be run only from Xcode.

Also adds initialization tests to docker, since they've been refactored a while ago and can now be called along with the other tests.