Test: change anonname keys into a static string
These keys are set to seven digit hex numbers that seemingly change with every execution, which can be annoying when running the bootstrapped tests if they update the AST dumps in a different execution than the one in which their comparison tests were updated. This should solve that problem without losing any valuable information (since the anonname key wasn't used for anything in any case).