Test: soften the issues with the lack of the Swift toolchain
This remove the test failure when the toolchain hasn't been detected and softens the language of the warnings. It should make it easier for contributors to understand what's happening if they don't have to track down failed Bootstrapping tests or find out why the alert emoji doesn't correspond to a failed test.