Meta: Update the Package.swift file and add metadata
Adds metadata on the supported Swift versions and macOS versions (Linux isn't currently supported in the Package metadata).
The `swiftLanguageVersions` property only supports `.v5`, not `.v5_1` or `.v5_2`.
The platform is set to macOS 10.13 since that's the earliest we're using (in the Shell.swift file). The 10.13 here corresponds to the minimum supported version.
Also change the Swift tools version to 5.1 because 4.0 didn't support the `swiftLanguageVersions` property.