Bugfix: Fix merge error
CallExpressions don't have types when used as TemplateExpressions, which can be a problem because the type is needed to create this DeclarationReferenceExpression. Making Calls have types can't be done because of templates, and making Declarations have optional types seems like it would be a lot of work. This hack is being used instead, for now, in the hopes that it won't cause problems.