Add capacity to transform a SwiftSyntax in a PrintableTree
The SwiftSyntax structure seems too complex to directly conform to PrintableAsTree. Instead, this class visits each node and turns it into a PrintableTree. The class is basically a copy/paste of SwiftSyntax's SyntaxAnyVisitor class, but including the names of the nodes during the visiting process (and with the logic for building a PrintableTree).