Support single-expression functions.
This change removes curly braces and instead replaces them with an equal sign.

Due to changes in Swift 5.1, functions with a single expression and a return type
can omit the 'return' keyword. The feature is heavily used in SwiftUI, for example.
According to the Swift manual (for Swift 5.3, https://docs.swift.org/swift-book/ReferenceManual/Expressions.html):

>  A closure that consists of only a single expression is understood to return the value of that expression.

Kotlin calls these "single-expression functions". See the documentation for more
info: https://kotlinlang.org/docs/reference/functions.html#single-expression-functions.

With this change, the following translation is done:

    func example() -> String {
      "Hello, World"
    }

    fun example(): String = "Hello, World"

The translation is only done when the function has a return type that is
equal to the value of the expression.

See issue #72.