Bugfix: use a bash script to call ruby so we can set the gem path
The current method of installing Gryphon through homebrew was failing to install Xcodeproj.

Homebrew can't install Xcodeproj in its default folder, since that requires sudo. Instead we have to:
- Install it in Gryphon's homebrew folder;
- Tell Gryphon where that folder is (by changing the contents of the new RubyScriptContents.swift file);
- Build and install Gryphon with the new (altered) source code.
- Make sure Gryphon uses the new bash script (generated on `gryphon init`) instead of calling ruby directly.

Basically, the RubyScriptContents.swift file will contain the contents of a bash script file that calls ruby, forwarding any arguments it has. That bash file will be created alongside other scripts when calling `gryphon init`. This is useful because the Homebrew formula can alter the Swift file to make Gryphon generate a bash file that also sets the `$GEM_PATH` variable to the place where Xcodeproj was installed before calling ruby.

This commit implements Gryphon's side of the fix; the rest of the code has to be done on the Homebrew formula's side.

Also improve the error message when Xcodeproj is not found, linking to the GitHub page with instructions for installing it.