Match implicit type expressions with explicit dot expressions
When the type is spelled out, Swift interprets it as a DotExpression (e.g. `A.B.c`), but when it's implicit it's a TypeExpression (e.g. an implicit `A.B` before a `.c`). This change makes these two forms match each other. The dot expressions can be of any length (>= 1).