Support matches with implicit `self` using SourceKit's indexing
The implicit `self` when referencing properties isn't spelled out, so SwiftSyntax doesn't have information on it (nor SourceKit's ExpressionTypes). This means we can't include it in the AST, so it doesn't match the right templates. For instance, `_string.startIndex` won't match `startIndex` in an extension of String if the `self.` is implicit.

This is fixed by using SourceKit's Indexing request, which associates expressions with USRs that include information on their parent type. We can run through all Indexing requests looking for type USRs, log them all in a dictionary, then match them to the USRs in expressions when needed to know what is the parent type.

Still, we don't know if there _was_ an implicit `self.` or not. The matching algorithm tries to find a match without it, then looks into the indexing information if no matches were found.