Avoid function translations for inexistent struct initializers
Swift creates an implicit memberwise initializers for structs that don't declare any other initializers. In these cases, Gryphon creates an additional function translation to match that implicit initializer. However, if there is an initializer that's being ignored with a translation comment, Gryphon would create the function translation even though it shouldn't. This is now fixed.