Meta: specify ff-only when pulling the bootstrap branch
This avoids git raising a warning (when pulling gryphon-old) saying we didn't specify if we wanted to merge or rebase the pulled changes. Since it should be rare to have committed changes in this folder, we set ff-only, which fails when there are any commits and makes us decide what to do on a case-by-case basis.