Bugfix: support float literals as autoclosures
Also improve error messages when float literals can't be converted, and turn some of those errors into warnings. Cases that raise warnings will be translated as Doubles (instead of Floats) since there should be literals that can be represented as Doubles but not as Floats (so using Double covers all cases, but using Floats woudn't).

Resolves #100.