Feature: support swiftASTDump files as input
This allows the unit tests to run using the Swift 5.5 AST dumps, which is much easier than implementing Swift 5.6 (and later) support for this old Gryphon version.

Also updates the tests to always use the Swift 5.5 AST Dumps, which will need to be added to the repo.

The main branch will need to be updated to run the bootstrap tests using the Swift 5.5 AST dumps as well.